Prevent relay mining (#77)
Prevent relay mining
