+++ validation/proxy/api_server/asgi.py
@@ -1,4 +1,4 @@
-from fastapi import FastAPI, Request, Response
+from fastapi import FastAPI, Request
 from fastapi.responses import JSONResponse
 from starlette import status
 import uvicorn
