+++ validation/proxy/validation_utils.py
@@ -1,19 +1,14 @@
 import sys
-import time
 import traceback
 from typing import Optional
-from typing import Type
 from typing import Union
 from PIL import Image
 from rich.console import Console
 import bittensor as bt
 import fastapi
-import httpx
-from config.validator_config import config as validator_config
 from fastapi import HTTPException
-from pydantic import BaseModel
 import random
-from core import constants as core_cst, utils as core_utils
+from core import utils as core_utils
 from models import utility_models, base_models
 import numpy as np
 
@@ -30,16 +25,6 @@ def log_task_exception(task):
         bt.logging.error(f"Exception occurred: \n{tb_text}")
 
 
-def get_synapse_from_body(
-    body: BaseModel,
-    synapse_model: Type[bt.Synapse],
-) -> bt.Synapse:
-    body_dict = body.dict()
-    body_dict["seed"] = random.randint(1, core_cst.LARGEST_SEED)
-    synapse = synapse_model(**body_dict)
-    return synapse
-
-
 def handle_bad_result(result: Optional[Union[utility_models.QueryResult, str]]) -> None:
     if not isinstance(result, utility_models.QueryResult):
         message = "I'm sorry, no valid response was possible from the miners :/"
@@ -49,43 +34,6 @@ def handle_bad_result(result: Optional[Union[utility_models.QueryResult, str]]) 
             status_code=fastapi.status.HTTP_400_BAD_REQUEST,
             detail=message,
         )
-
-
-def health_check(base_url):
-    try:
-        response = httpx.get(base_url)
-        return response.status_code == 200
-    except httpx.RequestError:
-        print(f"Health check failed for now - can't connect to {base_url}.")
-        return False
-
-
-def connect_to_external_server() -> str:
-    hotkey_name = validator_config.hotkey_name
-
-    servers = {
-        core_cst.EXTERNAL_SERVER_ADDRESS_PARAM: validator_config.external_server_url,
-    }
-
-    # Check each server
-    for name, url in servers.items():
-        if url is None:
-            raise Exception(f"{hotkey_name}.{name.upper()} not set in the config")
-
-        retry_interval = 2
-        while True:
-            connected = health_check(url)
-            if connected:
-                bt.logging.info(f"Health check successful - connected to {name} at {url}.")
-                break
-            else:
-                bt.logging.info(
-                    f"{name} at url {url} not reachable just yet- it's probably still starting. Sleeping for {retry_interval} second(s) before retrying."
-                )
-                time.sleep(retry_interval)
-                retry_interval += 5
-                if retry_interval > 15:
-                    retry_interval = 15
 
 
 def alter_image(
